+++ .gitignore
@@ -6,4 +6,4 @@
 .pdm-python
 build/
 __pycache__
-/nonebot_plugin_picstatus/res/picstatus-debug.html
+/nonebot_plugin_picstatus_ng/res/picstatus-debug.html
