+++ nonebot_plugin_picstatus_ng/collectors/bot.py
