+++ nonebot_plugin_picstatus_ng/collectors/cpu.py
