+++ nonebot_plugin_picstatus_ng/config.py
