+++ nonebot_plugin_picstatus_ng/misc_statistics.py
