+++ nonebot_plugin_picstatus_ng/res/assets/default_bg.webp
