+++ nonebot_plugin_picstatus_ng/res/js/globalThis.d.ts
