+++ nonebot_plugin_picstatus_ng/res/js/init-global.js
