+++ nonebot_plugin_picstatus_ng/res/js/load-plugin.js
