+++ nonebot_plugin_picstatus_ng/templates/__init__.py
