+++ nonebot_plugin_picstatus_ng/templates/default/__init__.py
