+++ nonebot_plugin_picstatus_ng/templates/default/res/css/no-shadow.css
