+++ nonebot_plugin_picstatus_ng/templates/default/res/css/theme-dark.css
