+++ nonebot_plugin_picstatus_ng/templates/default/res/templates/index.html.jinja
