+++ nonebot_plugin_picstatus_ng/templates/default/res/templates/macros.html.jinja
