+++ pyproject.toml
@@ -33,5 +33,8 @@ homepage = "https://github.com/wyf9/nonebot-plugin-picstatus-ng"
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
+[tool.hatch.build.targets.wheel]
+packages = ["nonebot_plugin_picstatus_ng"]
+
 [tool.hatch.version]
-path = "nonebot_plugin_picstatus/__init__.py"
+path = "nonebot_plugin_picstatus_ng/__init__.py"
